Fix failing unit tests
We are not sure why this wasn't a problem before (possibly due to
updated module dependencies), but we are seeing two upload tests fail
with "anachronistic" (not yet valid) authentication tokens, which we
believe to be due to having generated the token with the `pbench_token`
fixture outside of the time freeze context.

Fixed by dropping the time freeze in one test where it wasn't necessary,
and generating the authentication token for the other in the past rather
than outside the time freeze.