Fix failing unit tests
I'm not sure why this wasn't a problem before, but I was seeing two upload
tests fail with "anachronistic" (not yet valid) authentication tokens, which
I believe to be due to having generated the token with the `pbench_token`
fixture outside of the time freeze context. Fixed by dropping the time freeze
in one test where it wasn't necessary, and generating the authentication token
for the other in the past rather than outside the time freeze.