Remove file system state directories (#2994)
* Remove file system state directories

PBENCH-577

The Pbench Server has traditionally maintained controller-specific directories
containing symbolic links to tarballs in order to drive the components of the
server pipeline (from intake to backup to unpacking to indexing and culling).
We need to move away from this to support S3 migration and eventual message
passing on-demand component orchestration.

This is a first step. This removes the on-disk symlinks using a shim class
that manipulates Dataset metadata, with the idea that later it can be
refactored to wait-on/send messages through a message broker, deleting the
temporary `server.operation` metadata key.

This also drops the "dispatch" component, which really didn't do much aside
from another redundant validation of tarball MD5, and the "quarantine" directory
and Dataset state.