Correct use of Python 3.9 for the Server
When using Python 3.9 on RHEL 8 and it derivatives, one must invoke
Python 3.9 via `/usr/bin/python3.9`, and `python3` is either not there
or potentially a symlink to `Python 3.6` (in one particular case because
PCP was installed first).

We need to use the correct version of Python both in the command and in
the `PYTHONPATH`, along with using the `lib64` site packages.