Support browser reload of React component URL (#3058)
* Support browser reload of React component URL

PBENCH-778

When viewing a Dashboard component page, React displays a pseudo URL path in
the browser location bar; however this URL, for example /dashboard/login, is
not a distinct file on the host system and can't be loaded.

This change introduces a "pbench-in-a-can" Apache config file change that will
cause Apache to "rewrite" all URLs under /dashboard to redirect to the React
App code. The React infrastructure takes care of launching the component.

Since we used to install a /etc/http/conf.d/pbench.conf file as part of the
Ansible server configuration, and we no longer have server Ansible setup, this
also adds an "example" HTTPD configuration file that can be copied during
installation. (Note that we probably don't want to automatically install this
in the RPM spec file as it won't be used with an external NGINX, but that will
require similar mechanism. Which I'm willing to include in the sample NGINX
config if someone can tell me how that would look.)

Note that the `<Directory>` spec here is taken mostly intact from a web search
and may have excessive verbage. I haven't experimented too much, but any
improvements are welcome.