secure tage in cookies (#3063)
A cookie with the Secure attribute is only sent to the server with an encrypted request over the HTTPS protocol.

As our server is of HTTP protocol, the cookies are not being set. These cookies are used to get the user login details.

With KeyCloak integration, the entire mechanism for maintain the user login details will be changed so there will be no need to maintain these cookies.