Avoid ConfigParser interpolation (#3074)
PBENCH-978

A dataset may be gathered on a host with a scoped IPv6 address (represented in a
format like 0::0%eth0). The default `ConfigParser` used to read the dataset
`metadata.log` interprets `%` characters as a substitution of another config
value, and raises an exception because the "eth0" key doesn't exist.

Any `ConfigParser` instance used to process `metadata.log` rather than actual
config files chould be created with `interpolation=None`.