PUT functional test (#3064)
PBENCH-858

This is a preliminary (happy path) functional test for dataset upload using
PUT. I grabbed the tarballs and MD5 files from the legacy test cases Webb
had previously identified for his "with data" container.

It turns out not all of these are "good", and I segregated a number of these
into a "bad" subdirectory where we can investigate using them for failure
tests. (There may be a couple here that are technically "good", too; there's
a possible problem with duplicate uploads, and I re-used the container perhaps
a few too many times.)

This new functional test uses PUT to upload a set of datasets from a staging
directory, then uses `datasets/list` to compare the reported datasets. It then
captures the full list of datasets and waits for each to get to Indexed state.