Reduce datasets for functional test
While it's hard to diagnose server failures in the current limited functional
test environment (without access to logs), the tests are failing here because
we don't reach "backup" state, which I have seen happen as a result of filling
the filesystem.

I'm trimming off some of the redundant tarballs in an attempt to limit the space
requirements.