Remove use of unused `default-host`
Remove use of unused `default-host` and anything that depended on it.
This allows us to avoid fixing up the pbench server configuration file
for the in-a-can server container.

We also make a change to always copy the HTTPD pbench server conf file
into place, but only fix it up with the right host name for the in-a-can
environment.