Fix unpack sequencing problem (#3078)
* Fix unpack sequencing problem

PBENCH-975

The unpacker was attempting to re-unpack datasets, resulting in errors that
prevent indexing. The root problem ends up being that while `PUT` initially
enables each dataset for the `BACKUP` and `UNPACK` operations (which are
independent and can happen in parallel or in any order), the backup operation
was also (redundantly) enabling the archived dataset for `UNPACK`. In many
instances this works fine: however if the unpacker happens to run first, it
removed `UNPACK` from the enabled operations set and then the backup operation
*adds it back*, causing the operation to happen a second time. (And it's never
subsequently removed as the unpack operation never succeeds due to unexpected
`INCOMING` and `RESULTS` directories that already exist.)

This PR fixes the basic problem, and also adds some `try` blocks and error
logging to help catch unusual cases.