Delete API fails if dataset hasn't been indexed (#3080)
* Delete API fails if dataset hasn't been indexed

PBENCH-977

The `/datasets/delete` operation depends on the `server.index-map` metadata to
get a list of Elasticsearch document IDs. It failed if the map didn't exist.

Fix this so the wrapper cleanly detects that there are no bulk Elasticsearch
actions and skip the Elasticsearch operation, continuing to remove the SQL and
filesystem artifacts.

I also added a "safety" valve to reject deletion if something is mutating the
dataset by looking at the dataset's current state, so neither this operation
nor the other are confused.