API error improvements
PBENCH-995

The dashboard is sometimes receiving HTTP errors without `{"message"}` JSON
payloads. It's not entirely clear why this is happening given the `APIAbort`
logic, we could provide more useful hints in most cases and I've tried to
"fill in the blanks" in places where we'd omitted the `message` parameter.

In the course of this, I also added logic (mimicking that in several other
places) to aggregate errors when using `PUT /datasets/metadata` to set
multiple keys, when a weird error occurs in just a few. I realized that this
error path wasn't *tested*, and added a unit test.