Move Pbench Server to Listen on port 8080
By moving to use port 8080 in the Pbench Server container we allow for
the case of the container deployed by a non-root user using host
networking.

In order to make this work inside the Pbench-in-a-Can pod, we must move
the KeyCloak server to port 8090.  Within a pod, all containers are
executing within the same network namespace.  So we can't have Apache2
(httpd) and KeyCloak listening on port 8080, while only trying to map
the KeyClock port to 8090.

Further, there is only one authentication server URL.  The code does not
need to detect a fallback, that is up to the entity depoying the Pbench
Server to provide correctly in the configuration file.  The default use
for the Pbench-in-a-Can environment is also updated.

It should be noted that `hostPort` is required to expose a port outside
of the pod even if `containerPort` has the same value.

We also remove unnecessary host port specifications, default KeyCloak
server host name like we do for PostgreSQL, and wait for the Pbench
Server pod to come up for 10 minutes in the CI pipline to provide a bit
more time to debug.