Fix duplicate template error
PBENCH-450

Elasticsearch templates are loaded into SQL dynamically. There's a `find` to
determine whether each template already exists in the database; however during
startup it's possible for two independent tasks (e.g., indexer and backup) to
do this at the same time. Both might find that the template doesn't exist, and
attempt to create it. This sometimes results in a hang condition.

The duplication is not really a problem as both tasks would be loading the
same templates. Instead of failing, simply issue a `warning` log (to help
detect when this happens) and continue.