Promote access parameter to query param
PBENCH-1020

An overdue API cleanup I'd like to get in before the dashboard implements a
"Publish" control. This just moves the "access" parameter from the JSON
request body up to a URI query parameter. This will make it visible in the
HTTP/gunicorn log and makes it more consistent with our newer APIs.