Refactoring
I'm now reporting a "canned" client response message giving away nothing but
the generated UUID, while the internal log message can contain a more detailed
description of the error.

I moved generation of the log message from the constructor (which was weird),
to the _dispatch outer block.

The "partial success" cases in `/datasets/delete`, `/datasets/publish`, and in
`PUT /datasets/metadata` are now *success* cases with details in the JSON
response payload. I'm not 100% happy with this, either, but with the shift to
content-free client responses on 500, this seems the new least worst option.