Move expired token deletion off validation path
The process of validating an API call should NOT incur database deletion
traffic, where expired tokens could cause unwanted load on the database.
Instead, only during a new login are expired tokens purged from the
database.

The `created` column of `ActiveTokens` is replaced by the new
`expiration` column to make it easy to find all expired tokens for a
user and remove them.

Further, we add a SQLAlchemy `relationship()` between the active token
table and the user table so that it is trivial to fetch the `User` from
the `ActiveTokens` object.

The operation of adding a token for a user is moved to its own method,
`add_token()`.