API error improvements (#3091)
* API error improvements

PBENCH-995
PBENCH-459

The dashboard is sometimes receiving HTTP errors without `{"message"}` JSON
payloads. It's not entirely clear why this is happening given the `APIAbort`
logic, we could provide more useful hints in most cases and I've tried to
"fill in the blanks" in places where we'd omitted the `message` parameter.

A) Limit internal system error messages to a minimum level of detail while
B) Add a UUID to the client response message and system log, supporting analysis

I'm now reporting a "canned" client response message giving away nothing but
the generated UUID, while the internal log message can contain a more detailed
description of the error.

The "partial success" cases in `/datasets/delete`, `/datasets/publish`, and in
`PUT /datasets/metadata` are now *success* cases with details in the JSON
response payload. I'm not 100% happy with this, either, but with the shift to
content-free client responses on 500, this seems the new least worst option.