Use one pod.yml file; two images; func test script
We now have one `pod.yml` file instead of two with almost identical
contents. This necessitates moving the CI pipeline image names to
`jenkins/Pipeline.gy`.

We now have two images we build: (new) a stand-alone Pbench Server
image, and a the original "Pbench-in-a-Can" image used by the CI.
The stand-alone image is built first, and then used by the CI image
build to construct that iamge.

Finally, we provide a wrapper script, `jenkins/run-server-func-tests`,
for running server functional tests to allow for sane defaults to be
chosen (`envsubst` does not allow for that).  The new script handles
properly waiting for the pod to show up for a maximum of 10 minutes
before declaring failure.

Tested locally with another convenience script that we add,
`jenkins/runlocal`.