Allow setting dataset access on upload (#3116)
* Allow setting dataset access on upload

PBENCH-1029

The 0.69 "pass-through" compatibility server will be uploading datasets to
the 1.0 server under a fixed username, which means these datasets won't be
visible to other users.

This adds an "access" query parameter to `PUT` that allows the put-shim
server to upload all datasets as "public" so that they're visible to be
analyzed and claimed.

(NOTE that this implies some follow-on Agent work in pbench-results-move
and supporting classes to allow put-shim to specify the new query
parameter.)

To "easily" accomplish this, I refactored the upload class on top of our
`ApiBase` to handle parameter processing. Some changes were necessary to
avoid consuming the entire data stream during validation, but this also
allows `PUT` to participate in the new `APIInternalError` UUID mechanism
cleanup. `PUT` continues to handle auditing itself as we have no resource
ID to log at the beginning of `_dispatch`.