Simplify user reg & dump progress in func tests
The register user fixture is promoted to "module" scope, since it only
needs to happen once. This avoids repeated API calls to create a user
for every test that uses the `login_user` fixture.

Further, we change the `login_user` fixture to logout the user when the
test finishes.

Finally, we run the functional tests so that we can see the progress of
the tests as they operate and execute.