Replace Apache2 with Nginx
We also move the document root for serving files to be:

    /srv/pbench/public_html

This has a number of advantages.  It gives us one external mount point
for the production deployment, so that the dashboard is served from the
same volume.  It also means the `404.html` and `50x.html` pages that
are referenced in the Nginx configuration can eventually provided by
the dashboard.

We can also take advantage of a `maintenance.json` "page", where the
presence of that file will cause Nginx to server a 597 code, and its
JSON payload will be pulled from that file.

Finally, this allows us to provide a method for serving the old legacy
pages for the Agent generated HTML.