Move running functional tests under `tox`
We can now use `tox` in a normal development environment run the
functional tests against a target server, no special container
environment necessary.  This is very helpful for development since one
can iterate much faster without the need of the container environment.

We have added a delete test to remove all the uploaded tar balls for the
`tester` user.

Finally, we have enhanced the tests to report the payload of failed APIs
so facilitate easier debugging.