Don't emit logs before `rollback()`
When handling exceptions, certain kinds of exceptions in either
SQLAlchemy or in PyscoPG2 will cause references to the SQLAlchemy object
instance to cause other exceptions.  For example, one might see:

    sqlalchemy.exc.PendingRollbackError: This Session's transaction has
    been rolled back due to a previous exception during flush. To begin
    a new transaction with this Session, first issue Session.rollback().

By moving the log messages after the `rollback()` call, we avoid the
extra mess caused.

We also use the `self.logger` where possible, and we enhance the
existing state transition error log with the object information, and
add a debug log message on the success path.