Verify allowed methods before allowed operation
If an API is invoked using an unsupported method (e.g. `OPTION`), we
don't want to encounter a stack trace checking for a disabled API looked
up using the wrong method.

Further, we also need to ensure that we have a schema for the given
method before we check if the kind of operation is disabled.