Leverage context variables for config and logger
We add a new module, `pbench.server.globals`, which provides a
`ContextVar` called `server`.  The top-level interfaces for the server
(`shell`, `tree-manager`, `server/bin/*` commands, etc.) are responsible
for setting up the server context.

The server context consists of three pieces:

 * `config` - A PbenchServerConfig instance
 * `logger` - A reference to the logger instance in use
 * `db_session` - [Optional] the current database session in use

The majority of these changes are removing the `logger` fields from
objects, using `server.logger` instead.  A somewhat smaller set, though
still fairly large, are changes to remove embedded configuration object
references in objects.  And lastly, we promote the database "session" up
to use the same mechanism instead of via a class variable.

The biggest reason to make this change is that by nature and design,
loggers are singletons.  Within the context of a python interpreter
instance, each logger has a name, and is fetchable in any context by
using `logging.getLogger('<name>')`.  As you can see from these changes,
by removing the need to pass a logger around, we reduce a substantial
amount of code "mechanism."  By embracing the nature of singleton
loggers we avoid the need to ensure loggers are available in all
contexts referenced through object instances.

Arguably the second biggest reason is to leverage the nature of context
variables so as to facilitate removal of code mechanism for other
singletons, the configuration in use and the database session in use.

A third reason is that it also helps prepare for future concurrency
mechanims ("async", which is not really asynchronous, and / or threads).

The motivating reason for these changes is to ensure we have a handle on
what loggers are in use for changes to how those loggers behave (adding
/ removing different handlers, in particular).

Since most of the code for the server leverages Flask which already uses
context variables (see `current_app` and the `g` object), we could have
just used those existing mechanisms.  We chose to create our own so that
we can consider contexts where we don't have, or don't want, Flask
involved (yes, we are aware that Flask has support for CLIs).

For the server unit tests, we now use a session-scoped fixture to create
a logger using "brace" formatting.