Replace Apache2 with Nginx
Along with this change, we add rsyslog for observability, where we have
Nginx sent all logs to syslog directly, using a JSON format with the
`@cee:` prefix to allow rsyslog's `mmjsonparse` module to recognize the
JSON and add those fields to the rsyslog message (another rsyslog
indexer handles adding those fields to an OpenSearch or Elasticsearch
instance).

Since the OIDC server work is not complete, we change the code to allow
for the configuration to optionally specify an authentication section
where the missing section means we won't try to use an OIDC server.

We move the document root for serving files to be:

    /srv/pbench/public_html

This has a number of advantages.  It gives us one external mount point
for the production deployment, so that the dashboard is served from the
same volume.  It also means the `404.html` and `50x.html` pages that
are referenced in the Nginx configuration can eventually provided by
the dashboard.

We can also take advantage of a `maintenance.json` "page", where the
presence of that file will cause Nginx to server a 597 code, and its
JSON payload will be pulled from that file.

The Nginx configuration also allows us to provide a method for serving
the old legacy pages for the Agent generated HTML.

Remove use of unused `default-host` and anything that depended on it.
This allows us to avoid fixing up the pbench server configuration file
for the in-a-can server container.

We bring the `jenkins/runlocal` inline with CI pipeline, and add a more
informative comment for token logins.

Fixed a bug in the server functional tests script to not be so
aggressive and run cleanup on every `ERR`or of a shell command.

Finally, we remove a redundant `PB_POD_NAME` setting from the CI
pipeline.