Fix PUT metadata partial success
PBENCH-1049

During the course of making an unrelated change, I "broke" a metadata test
case for the "partial success" case of `PUT /datasets/metadata`. The test was
still expecting the previous 500/internal-server-error behavior even though
this had been changed to a success with an advisory message.

The reason the test "succeeded" was that the product code attempted to report
the issue with an error log message, which failed because it involved trying
to serialize an `Audit` record via `json.dumps` which apparently is unhappy
with the `Audit` (and probably all SQLAlchemy proxy) objects, and the failure
caused the API call to terminate with ... an internal server error.

This PR fixes the test and some aspects of the implementation, including
making all metadata exceptions extend `MetadataError` to provide an easy
catch-all handle.