Move running functional tests under `tox`
We can now use `tox` in a normal development environment run the
functional tests against a target server, no special container
environment necessary.  This is very helpful for development since one
can iterate much faster without the need of the container environment.

We have added a delete test to remove all the uploaded tar balls for the
`tester` user.

We refactored to the tests to make them operationally independent.
We drop the `self` parameter in favor of all `staticmethod` tests since
the `tarballs` class variable was not really helpful to share between
tests.  We also add the required test ordering in the doc strings.

Expanded the use of the print statements in the other `test_put.py`
tests for consistency (makes for a nice view of what was uploaded,
indexed, and then deleted).

Finally, we have enhanced the tests to report the payload of failed APIs
so facilitate easier debugging.