Apply corrections for when tests are skipped
The `Commons` class provides a set of shared tests as a "base" class, of
sorts, for various other tests classes to use to help validate
behaviors.

However, it wasn't clear what tests were supposed to be skipped or not.

**_It is this author's opinion that we should drop test skipping in the
future._**

In the mean time, we have applied some comments, reworked existing
skip conditions, added other conditions, and removed unused skips
in an attempt to verify what is skipped is expected.

This actually resulted in an _increase_ in the number of skipped tests
due to the fact that one of the tests was not actually executing when
a parameter was missing.

Move comments about why a test is skipped into the skip message itself.