Don't use xdist plugin for serial pytests
If we use the `-n 1` pytest argument we still run using the `xdist`
plugin, but it just uses one process.  Since the `xdist` plugin changes
the test environment a bit, don't use it when trying to run tests one at
a time.

We also ensure that the optional positional arguments provided by the
caller come after all of the default pytest arguments.