Rework use of `get_pbench_logger` in unit tests
The unit tests use a fixed server configuration file which is generated
once per testing session (for all tests discovered).

Only one test suite, `test_user_management_cli`, requires the use of the
`_PBENCH_SERVER_CONFIG` environment variable for testing since that
module is actually testing a CLI interface.

We take advantage of this to make the `server_config` fixture scoped for
the "session", moving the use of `monkeypatch` to set the environment
variable to a fixture for `test_user_management_cli` suite.

With this change, we can use the one `server_config` fixture (removing
the other ones), and then also promote the `make_logger` fixture to be
"session" scoped as well.  This means we only create a "TEST" logger
once for all tests for the session and then use it were needed.

As a result we can remove the use of `get_pbench_logger` from all the
other test suite files.