Refactor `PbenchServerConfig`
This is a major refactoring of the `PbenchServerConfig` class and top
level server module.

The first change is to remove unused properties for `server_config`,
`mail_recipients`, `_unittests`, and `_ref_datetime`.

Then we use properties for all the attributes set in the constructor
which require the use of instance methods.  Since the original behavior
was for the constructor to raise `BadConfig` for missing required
options, we create a module method, `get_pbench_server_config()` to
provide the same behavior.

We also take the opportunity to update and correct the doc strings for
the whole module.

As a consequence of removing the `_ref_datetime` property, we had to
update `pbench-cull-unpacked-tarballs`, and as a result of removing
`_unittests` we took the opportunity to remove the unused "mock" code
in the `pbench.server.s3backup` module.