Rework database modules and class names
The main goal for this work is consistency in naming surrounding the
database models.

Most of the code refers to "auth tokens" in various contexts. As such,
"active token(s)" names are renamed to "auth token(s)".

Related to that, we update other database modules so that they follow
the same pattern of singluar module name and plural table names.  The
plural table names are more natural when constructing SQL queries.  We
also change the database module model names to follow their table names.

One notable change is that the `database.model.server_configs` module
has been renamed to `database.model.server_settings`, with the ORM class
name renamed from `ServerConfig` to `ServerSetting`.

We are making both changes in this commit so that we can provide one
database migration for the changes.