Provide the "Initial tables" alembic migration
Since we already have been using the database for quite some time, we
create an initial migration which will take an empty database to the
database base schema version as of commit 6a764f154. That commit was the
most recent version deployed in Red Hat's staging environment.

In order to make it convenient to work with migrations, we also create a
`tox` environment, `alembic`, which simply invokes the new script
`alembic.run` which one can modify to execute whatever Alembic commands
are needed.  The committed copy takes no actions, but has the initial
commands used to create the first migration:

    alembic stamp head
    alembic revision --autogenerate -m "Initial tables"

We also update the `alembic.ini` file to assume a `localhost` PostgreSQL
database, and provide a Pbench Server configuration file that works in
the Alembic environment.