Integrate `alembic check` into basic server tests
Starting with Alembic v1.9.0 the command, `alembic check` is available
which will succeed if no migrations are required, and fail if it detects
that a migration is necessary.

We also remove the unused `alembic/alembic.ini`, drop the need for a
separate `tox` environment, remove the need for a `pbench-server.cfg`
file entirely by reverting to the original env.py as generated by
`alembic init alembic`.

We also collapse the existing migrations to one since we just need one
to move from an empty database to our current schema.