Provide the "Initial database schema" alembic migration (#3174)
Since we already have been using the database for quite some time, we
create an initial migration which will take an empty database to the
database base schema version as of commit 6a764f154.  That commit was
the most recent version deployed in Red Hat's staging environment.  The
existing revision migration was removed so that we only have a migration
from empty database to fully populated.

Starting with Alembic v1.9.0, the command, `alembic check`, is available
which will succeed if no migrations are required, and fail if it detects
that a migration is necessary.  So future PRs will have to provide a
migration revision if a change to the database schema is made in the PR,
where `alembic check` will fail if one is not provided.

A new `tox` environment, `alembic-check`, will now run the `alembic
check`.  Invoked via:

    tox -e alembic-check

The actual check is provided by the script:

    jenkins/run-alembic-migrations-check

That script runs a very simple PostgreSQL container against on the
`localhost` (which is why `alembic.ini` is configured to reference a
`localhost` database).  It then runs all our known migrations (`alembic
upgrade head`) and then ask Alembic to check that we don't have any
additional model changes which need a migration (`alembic check`).

The CI build will fail when any migrations are missing, and a local
`tox -e alembic-check` will also fail.

We have updated the `alembic.ini` file with the latest version generated
by alembic 1.9.x.

The previous requirement of needing a `pbench-server.cfg` file has been
removed by restoring the `alembic/env.py` file back to the default with
a small change to reference our database metadata.

We have removed the unneeded `alembic.ini` file in the `alembic` sub-
directory.