Change sync-satellite to use reception area
PBENCH-594

This turned out to be a bit more subtle, but in the end not too bad.

Basically, we just want `pbench-sync-satellite` to create the same final
state that `pbench-copy-result-tb` creates, which is just placing a .md5
and .tar.xz in a controller directory in the receive tree.  But we need
to maintain the remote updates for the satellite.

However, we have some legacy vestiges of the very old `.prefix` files.
Rather than try to carry over that supporting behavior from sync-
satellite, I just removed it entirely from the rest of the code base.

The unit tests proved very useful to see the behaviors that were taking
place, and they have been updated accordingly. Note that the .prefix
files are ignored and show up in the test output either as unexpected
files and directories and as tar balls that no longer have a prefix.

Co-authored-by: Peter Portante <peter.portante@redhat.com>