Fix repeated indexing
Fixes issue #3198.

The root cause was that we have operation names for indexing which all
had "INDEX" in the name: "INDEX", "INDEX_TOOL", and "RE_INDEX".

If any dataset had in its list of operations either "INDEX_TOOL" or
"RE_INDEX", then any search for "INDEX" would also match those option
names because a `contains()` operation was used in the query filter.  If
all the operation names did not have any overlaps, that would work.

The fix was to filter locally by the exact operation name.  Since the
total number of datasets on the system are rarely in the same
operational state, it seems like a point solution that solves the
problem efficiently without to many cascading changes.

Also includes:

 * Removed `COPY_SOS` op, added `Sync` class unit test
   * The copy sos reports functionality was remove long ago.
   * We added a unit test to demonstrate that a dataset with an
     "operation" of "INDEX_TOOL" is NOT found when looking for "INDEX".
 * Removed `maintenance` role; added verify backup step