Build and run functional tests all locally
The `jenkins/runlocal` script now builds the RHEL 9 RPM from the
checked-out sources in the current working tree, builds the Pbench
Server CI container, and runs the functional tests against the CI
container image.

As a result, `jenkins/runlocal` no longer pushes images to a
registry.

We modify the Pbench-in-a-Can `pod.yml` file to accept a one
environment variable for the full image name to use, moving the logic
of constructing that name into `run-pbench-in-a-can`.